+++ src/modules/meal/meal.dto.ts
@@ -6,7 +6,7 @@ export class CreateMealDto {
     readonly author: string;
 
     @IsNotEmpty({ message: 'Meal should have a description' })
-    @Length(3, 16)
+    @Length(3, 256)
     readonly description: string;
 
     @IsOptional()