+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.12.0] - 2023-09-05
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): REMOVE RECORD >>>> Basic tests for `AuthService`.
+- [Marek Kurańda](https://github.com/mjkuranda): Unit tests for `UserService`.
+- [Marek Kurańda](https://github.com/mjkuranda): Unit tests for `MealService`.
+- [Marek Kurańda](https://github.com/mjkuranda): Unit tests for `IngredientService`.
+- [Marek Kurańda](https://github.com/mjkuranda): Unit tests for `AuthService`.
+
+### Changed
+- [Marek Kurańda](https://github.com/mjkuranda): `space-before-function-parens` allows space after `async` keyword.
+
 ## [1.11.0] - 2023-09-04
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Logger messages to the others modules.